// lib/types/threejs.type.ts
export type AndroidType = GLTF & {
  nodes: {
    AF_Sensor_Glass: THREE.Mesh;
    Antenna_Plastic: THREE.Mesh;
    Back_Cam_Deco: THREE.Mesh;
    Back_Cam_Deco_Black: THREE.Mesh;
    Back_Cover_Glass: THREE.Mesh;
    Bezel: THREE.Mesh;
    Cam_Bezel: THREE.Mesh;
    Cam_Body: THREE.Mesh;
    Cam_Glass: THREE.Mesh;
    Cam_Lens: THREE.Mesh;
    Cam_Mt: THREE.Mesh;
    Display_ActiveArea: THREE.Mesh;
    Flash: THREE.Mesh;
    Flash_Glass: THREE.Mesh;
    Rear_Case: THREE.Mesh;
    Rearcase_Hole: THREE.Mesh;
    S_Pen_Tip: THREE.Mesh;
    Samsung_Logo: THREE.Mesh;
    Usb_1: THREE.Mesh;
    Usb_2: THREE.Mesh;
    Zoom_Cam: THREE.Mesh;
    Display_ActiveArea001: THREE.Mesh;
  };
  materials: {
    AF_Sensor_Glass: THREE.MeshStandardMaterial;
    Antenna_Plastic: THREE.MeshStandardMaterial;
    Back_Cam_Deco: THREE.MeshStandardMaterial;
    Back_Cam_Deco_Black: THREE.MeshStandardMaterial;
    Back_Cover_Glass: THREE.MeshStandardMaterial;
    Bezel: THREE.MeshStandardMaterial;
    Cam_Bezel: THREE.MeshStandardMaterial;
    Cam_Body: THREE.MeshStandardMaterial;
    Cam_Glass: THREE.MeshStandardMaterial;
    Cam_Lens: THREE.MeshStandardMaterial;
    Black_Hole: THREE.MeshStandardMaterial;
    main_glass: THREE.MeshPhysicalMaterial;
    Flash: THREE.MeshStandardMaterial;
    Flash_Glass: THREE.MeshStandardMaterial;
    Rearcase: THREE.MeshStandardMaterial;
    Rearcase_Hole: THREE.MeshStandardMaterial;
    Samsung_Logo: THREE.MeshStandardMaterial;
    Usb_1: THREE.MeshStandardMaterial;
    Usb_2: THREE.MeshStandardMaterial;
    Zoom_Cam: THREE.MeshStandardMaterial;
    ["main_screen.001"]: THREE.MeshStandardMaterial;
  };
};